+++ src/internal/text/TextLabel1.tsx
@@ -7,7 +7,7 @@ type TextLabel1React = {
 /* istanbul ignore next */
 export function TextLabel1({ children }: TextLabel1React) {
   return (
-    <span className='text-[#0A0B0D] text-bold text-sans text-sm leading-5'>
+    <span className="text-[#0A0B0D] text-bold text-sans text-sm leading-5">
       {children}
     </span>
   );
