--- src/swap/utils.test.ts
@@ -1,121 +1,0 @@
-import {
-  formatTokenAmount,
-  getRoundedAmount,
-  isSwapError,
-  isValidAmount,
-} from './utils'; // Adjust the import path as needed
-
-describe('isValidAmount', () => {
-  it('should return true for an empty string', () => {
-    expect(isValidAmount('')).toBe(true);
-  });
-
-  it('should return true for a valid number string with a decimal point', () => {
-    expect(isValidAmount('123.45')).toBe(true);
-  });
-
-  it('should return false for a string with multiple decimal points', () => {
-    expect(isValidAmount('123.45.67')).toBe(false);
-  });
-
-  it('should return false for a string with non-numeric characters', () => {
-    expect(isValidAmount('123a45')).toBe(false);
-  });
-
-  it('should return true for a valid number string with no integer part but a decimal point', () => {
-    expect(isValidAmount('.12345')).toBe(true);
-  });
-
-  it('should return true for a valid number string with no decimal part', () => {
-    expect(isValidAmount('12345.')).toBe(true);
-  });
-
-  it('should return false for a string with spaces', () => {
-    expect(isValidAmount('123 45')).toBe(false);
-  });
-});
-
-describe('isSwapError function', () => {
-  it('returns true for a valid SwapError object', () => {
-    const response = {
-      error: 'Swap failed',
-      details: 'Insufficient balance',
-    };
-
-    expect(isSwapError(response)).toBe(true);
-  });
-
-  it('returns false for null or non-object inputs', () => {
-    expect(isSwapError(null)).toBe(false);
-    expect(isSwapError(undefined)).toBe(false);
-    expect(isSwapError('error')).toBe(false);
-    expect(isSwapError(123)).toBe(false);
-  });
-
-  it('returns false for objects without the "error" property', () => {
-    const response = {
-      message: 'An error occurred',
-    };
-
-    expect(isSwapError(response)).toBe(false);
-  });
-
-  it('returns false for empty objects', () => {
-    const response = {};
-
-    expect(isSwapError(response)).toBe(false);
-  });
-});
-
-describe('formatTokenAmount', () => {
-  test('formats amount correctly with 18 decimals', () => {
-    const amount = '100000000000000000';
-    const decimals = 18;
-    const formattedAmount = formatTokenAmount(amount, decimals);
-    expect(formattedAmount).toBe('0.1');
-  });
-
-  test('formats amount correctly with different decimals', () => {
-    const amount = '1000000000';
-    const decimals = 9;
-    const formattedAmount = formatTokenAmount(amount, decimals);
-    expect(formattedAmount).toBe('1');
-  });
-
-  test('handles very small amounts correctly', () => {
-    const amount = '1';
-    const decimals = 18;
-    const formattedAmount = formatTokenAmount(amount, decimals);
-    expect(formattedAmount).toBe('1e-18');
-  });
-
-  test('handles zero amount correctly', () => {
-    const amount = '0';
-    const decimals = 18;
-    const formattedAmount = formatTokenAmount(amount, decimals);
-    expect(formattedAmount).toBe('0');
-  });
-
-  test('handles large amounts correctly', () => {
-    const amount = '1000000000000000000000000000';
-    const decimals = 18;
-    const formattedAmount = formatTokenAmount(amount, decimals);
-    expect(formattedAmount).toBe('1000000000');
-  });
-});
-
-describe('getRoundedAmount', () => {
-  it('returns a rounded number with specified decimal places', () => {
-    const balance = '0.0002851826238227';
-    const fractionDigits = 5;
-    const result = getRoundedAmount(balance, fractionDigits);
-    expect(result).toBe('0.00029');
-  });
-
-  it('returns a rounded number with more decimal places than available', () => {
-    const balance = '123.456';
-    const fractionDigits = 10;
-    const result = getRoundedAmount(balance, fractionDigits);
-    expect(result).toBe('123.456');
-  });
-});
