+++ src/utils/getRoundedAmount.ts
@@ -1,0 +1,4 @@
+export function getRoundedAmount(balance: string, fractionDigits: number) {
+  const parsedBalance = Number.parseFloat(balance);
+  return Number(parsedBalance)?.toFixed(fractionDigits).replace(/0+$/, '');
+}
